api/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
alembic>=1.13
asyncpg>=0.29
pydantic>=2.5
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
httpx>=0.26
websockets>=12.0
